+++ backend/restobar-rdr/src/main/java/com/restobar1/restobar_rdr/controller/VentaController.java
@@ -28,7 +28,7 @@ public class VentaController {
     }
 
     @DeleteMapping("/{id}")
-    public void eliminar(@PathVariable Long id) {
+    public void eliminar(@PathVariable Integer id) {
         ventaService.eliminar(id);
     }
 }
