+++ backend/restobar-rdr/src/main/java/com/restobar1/restobar_rdr/services/CategoriaService.java
@@ -27,7 +27,7 @@ public class CategoriaService {
         return categoriaRepository.save(categoria);
     }
 
-    public Categoria editar(Long id, Categoria categoriaActualizada) {
+    public Categoria editar(Integer id, Categoria categoriaActualizada) {
         Categoria categoriaExistente = categoriaRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Categoría no encontrada con id: " + id));
 
@@ -36,7 +36,7 @@ public class CategoriaService {
         return categoriaRepository.save(categoriaExistente);
     }
 
-    public void  eliminar(Long id) {
+    public void  eliminar(Integer id) {
         boolean tieneProductos = productoRepository.existsByCategoria_Id(id);
 
         if (tieneProductos) {
